+++ NEWS.md
@@ -7,6 +7,9 @@
   * `Player.prototype.console`
   * `World.prototype.events`
   * `System.prototype.events`
+* Functions provided by `lib/delay.ts` now catch unhandled exceptions and
+  print it to `console`, if callbacks are actually async functions and they
+  throw.
 * Removed `lib/event-emitter.ts`.
 * Renamed `lib/enchantment.ts` to `lib/item/enchantment.ts`.
 * Renamed `lib/item-stack.ts` to `lib/item/stack.ts`.